+++ guidedmodules/models.py
@@ -32,12 +32,29 @@ class AppSource(models.Model):
         return self.slug
 
     def get_description(self):
+        # Return a description of the connection that this AppSource
+        # provides. The output should be safe to share with users of
+        # the site --- so credentials should not be returned in this
+        # string.
         if self.spec["type"] == "null":
             return "null source"
         if self.spec["type"] == "local":
             return "local filesystem at %s" % self.spec.get("path")
         if self.spec["type"] == "git":
-            return self.spec.get("url") + ("@"+self.spec["branch"] if self.spec.get("branch") else "")
+            try:
+                # Remove credentials from URL.
+                from urllib.parse import urlsplit, urlunsplit
+                (scheme, host, path, query, fragment) = urlsplit(self.spec["url"])
+                host = host.split("@", 1)[-1]
+
+                # Set fragment to git branch.
+                fragment = self.spec.get("branch")
+
+                # Re-form URL.
+                url = urlunsplit((scheme, host, path, query, fragment))
+            except Exception:
+                url = "<invalid url>"
+            return url + ("@"+self.spec["branch"] if self.spec.get("branch") else "")
         if self.spec["type"] == "github":
             return "github.com/%s" % self.spec.get("repo")
 
@@ -74,6 +91,39 @@ class AppInstance(models.Model):
         # For debugging.
         return "<AppInstance [%d] %s from %s>" % (self.id, self.appname, self.source)
 
+    def is_authoring_tool_enabled(self, user):
+        return (self.source.spec["type"] == "local" # so we can save to disk
+            and user.has_perm('guidedmodules.change_module'))
+
+    def has_upgrade_priv(self, user):
+        # Does a user have permission to ugprade the Modules in this AppInstance?
+        # Yes if the user is an admin of all the Projects that the Tasks that use
+        # the Modules are in. In practice, that's just the one Project that was
+        # created when this AppInstance was created.
+        #
+        # In the database, we allow AppInstances to be shared across many Projects.
+        # The system AppInstance which holds e.g. the user profile module *is*
+        # shared across many user projects, but that AppInstance is blacklisted
+        # from upgrades below. But in other cases there's an implicit constraint
+        # that Projects and AppInstances are one-to-one because when we "start"
+        # an app **we always create an AppInstance and a Project as a pair.**
+        #
+        # In an abundance of caution, the permission check is coded to not make
+        # that assuption that there is just one Project tied to this AppInstance,
+        # so we check that the user is an admin to all relevant Projects. We would
+        # not want to upgrade an AppInstance associated with a Task that the user
+        # does not have administrative rights to. If in the future the implicit
+        # constraint is removed, we would need to be more precise about what is
+        # upgrade since we would not want to upgrade all Projects at once necessarily.
+        if self.system_app: return False
+        projects = Task.objects.filter(module__app=self).values_list("project", flat=True).distinct()
+        if len(projects) == 0:
+            # This AppInstance doesn't appear to be in use! Well, lock it down.
+            return False
+        for project in projects:
+            if user not in Project.objects.get(id=project).get_admins():
+                return False
+        return True
 
 class Module(models.Model):
     source = models.ForeignKey(AppSource, related_name="modules", on_delete=models.CASCADE, help_text="The source of this module definition.")
@@ -216,10 +266,8 @@ class Module(models.Model):
             })
 
     def is_authoring_tool_enabled(self, user):
-        return (
-                self.app is not None # legacy Module not associated with an AppInstance
-            and self.source.spec["type"] == "local" # so we can save to disk
-            and user.has_perm('guidedmodules.change_module'))
+        # legacy Modules don't have self.app set
+        return self.app is not None and self.app.is_authoring_tool_enabled(user)
     def get_referenceable_modules(self):
         # Return the modules that can be referenced by this
         # one in YAML as an answer type. That's any Module
@@ -737,6 +785,8 @@ class Task(models.Model):
     def has_review_priv(self, user):
         return user in self.project.organization.reviewers.all()
 
+    def has_delete_priv(self, user):
+        return user in self.project.get_admins()
 
     # INVITATION TARGET FUNCTIONS