+++ guidedmodules/urls.py
@@ -22,6 +22,6 @@ urlpatterns = [
     url(r'^_authoring_tool/new-question$', guidedmodules.views.authoring_new_question),
     url(r'^_authoring_tool/edit-question$', guidedmodules.views.authoring_edit_question),
     url(r'^_authoring_tool/edit-module$', guidedmodules.views.authoring_edit_module),
-    url(r'^_authoring_tool/reload-app$', guidedmodules.views.authoring_edit_reload_app),
+    url(r'^_upgrade-app$', guidedmodules.views.upgrade_app),
 ]